Improved script that launches eclipse.

+++ usr_local_bin/eclipse.sh
@@ -2,7 +2,11 @@
 
 echo "Running eclipse ... "
 
-killall -9 eclipse
+IS_RUNNING=$(ps -AfH | grep "eclipse " | grep -v grep)
+if [ ! -z "${IS_RUNNING}" ] && [ "" != "${IS_RUNNING}" ]; then
+	echo "It is still running. Kill it! "
+	killall -9 eclipse
+fi
 
 # nc -zv 127.0.0.1 3128
 # if [ 0 -eq $? ]; then
@@ -16,7 +20,10 @@ killall -9 eclipse
 #echo "Proxy chosen: ${PROXY_IP} ${PROXY_PORT} "
 #echo " "
 
-ECLIPSE_PATH=$(find ~ -name eclipse -type f | head -n 1)
+ECLIPSE_PATH=$(find ~/local/ -name eclipse -type f | head -n 1)
+if [ -z "${ECLIPSE_PATH}" ] || [ "" == "${ECLIPSE_PATH}" ]; then
+	ECLIPSE_PATH=$(find ~ -name eclipse -type f | head -n 1)
+fi
 
 # ECLIPSE_OPTS="-Dhttp.proxyHost=${PROXY_IP} -Dhttp.proxyPort=${PROXY_PORT} -Dhttps.proxyHost=${PROXY_IP} -Dhttps.proxyPort=${PROXY_PORT} -DsocksProxyHost= -DsocksProxyPort=  -Djava.net.preferIPv4Stack=true -Djava.net.useSystemProxies=true -Dorg.eclipse.ecf.provider.filetransfer.excludeContributors=org.eclipse.ecf.provider.filetransfer.httpclient4 "
 # -Dhttp.proxyUser=<user>
